Force running Functest vs public endpoints
It sets OS_INTERFACE to force the use of public endpoints as the
internal and management endpoints are unreachable from jumphost.

It also sets OS_REGION_NAME to avoid the fallback mechanism in tempest
selecting the first endpoint listed in catalog (here internal) [1]

EXTERNAL_NETWORK is defined to allow running Functest testcases in
parallel.

[1] https://github.com/openstack/tempest/blob/21.0.0/tempest/lib/auth.py#L583

Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>
Change-Id: I9933316479d1830d3f37dcae19940be0c8f185b1