Update patch set 1
Patch Set 1:

Really appreciate the effort to update the tools to support multiple env, however, this approach seems harsh and requires redundant effort for each and every pod.   It seems we can pretty much re-use this yaml as a template, and simply populate the information we need via substitution by reading each site's networks.yaml file, which already contain info such as DNS servers and CIDR.  I look just now and the global network schema does supports config of ranges and gateway as well, so we can easily add that info to every site and then this saves us from having to incur this work over and over again, or if the template needs to change/update, it saves us from having to do it n times over.  As soon as I can find some time, I can help with this activity if we can agree this is a good path forward.  I am open to other suggestions as well.

Patch-set: 1