Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Please submit this patch for Iruya, thanks.

Iruya branch points to a wiki page for the installation guide. Do you see this commit (move from wiki to sphinx doc) as required for the Iruya release? Bin has already tagged Iruya release in Airship so I don't want to redo that if this patch is not a must for Iruya.

Patch-set: 5