Update patch set 1
Patch Set 1:

A couple of questions for Cedric. 1. the blacklist does seem to filter out the live migration test case. 2. the opnfv_nova.yaml change of net_id doesn't seem to address the glance load image and boot server failure. It still doesn't use the existing provider network. 3. the functest-smoke container run_test is running on python2.7 and the source code in 2.7 site-packages seems to be slightly different than 3.6.

Patch-set: 1