Update patch set 2
Patch Set 2:

The openstack.cred in /home/opnfv/functest is outdated. Updated it manually. But it'd be nice that it  come from the Airship site manifests so we can ensure it is always to be in sync with the deployed cloud. thoughts?

Patch-set: 2
Reviewer: Gerrit User 7590 <7590@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>